requirements.txt
# Core app
gradio>=6.9.0
llama-cpp-python>=0.3.20

# Offline Speech-to-text (Whisper via faster-whisper)
faster-whisper
torch==1.13.0
torchaudio==1.13.0

# Offline TTS (Tortoise TTS for human-like voice)
# tortoise-tts>=3.0.0
TTS>=0.21.0
transformers==4.56.2
simpleaudio 

# Audio handling
pydub

# Numerics
numpy

# Optional: if you ever use pandas in analytics/logging later
# pandas
